array indicies should be ints

+++ scripts/mrms/mrms_raster_p1h.py
@@ -55,8 +55,8 @@ def do(now, hr ):
         image = np.where( val < 0, 255, image)
         #print tile, np.min(image), np.max(image)
         ysz, xsz = np.shape(val)
-        x0 = (tilemeta['ul_lon'] - util.WEST) * 100.0
-        y0 = round((tilemeta['ll_lat'] - util.SOUTH) * 100.0,0)
+        x0 = int((tilemeta['ul_lon'] - util.WEST) * 100.0)
+        y0 = int(round((tilemeta['ll_lat'] - util.SOUTH) * 100.0,0))
         imgdata[y0:(y0+ysz),x0:(x0+xsz)] = val.astype('int')
 
     (tmpfp, tmpfn) = tempfile.mkstemp()
